backend/requirements.txt
psycopg2-binary
Django
djangorestframework
django-cors-headers

